requirements.txt
# ctransformers==0.2.5
# sentence-transformers==2.2.2
# pinecone-client
# langchain==0.0.225
# flask
# PyMuPDF

# Core libraries
langchain==0.3.27
langchain-core==0.3.78
langchain-community==0.3.31
langchain-pinecone==0.2.12
langchain-huggingface==0.3.1

# Embeddings + Models
sentence-transformers==2.2.2
ctransformers==0.2.5
transformers==4.36.0
# tokenizers==0.15.2
tokenizers==0.22.1

# Vector database
pinecone==6.0.1

# PDF + Web
PyMuPDF
flask

# Optional (if you use tqdm or numpy in your scripts)
tqdm
numpy

python-dotenv
-e .